refactored and created a custom hook useSortUsers
