+++ src/modules/UsersTable/types/User.ts
@@ -1,0 +1,11 @@
+export interface User {
+    _id: number,
+    firstName: string
+    lastName: string
+    email: string
+    phone: string
+}
+
+
+export type TableHeadTitles = Record<keyof User, string>
+
